requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
langchain-huggingface==0.0.3
chromadb==0.5.3
pypdf==4.2.0
sentence-transformers==3.0.1
pydantic==2.7.4
